refactor(AlarmController, MovieMainActivity): 알림 체널 생성 코드 이동
